+++ src/scene.js
@@ -628,6 +628,7 @@ class SpaceScene extends Scene {
           }
 
           if (a.collision(b)) {
+            window.sampler(`a2`, 0.3); // Asteroid impact
             const ae = a.e;
             a.e = b.e > 0 ? a.e - b.e : a.e; // Strange situations
             b.e -= ae;
@@ -651,7 +652,8 @@ class SpaceScene extends Scene {
             if (a.e < 0 && b.source === this.player._id) {
               this.player.increaseEnergy(settings.hull.pctRecoveredPerAsteroid);
               newAsteroids.push(...a.split(b.vel, this.debrisList));
-              window.sampler("a5", 1.9);
+              const r = Math.floor(Math.random() * 4);
+              window.sampler(`g${r}`, 1.9); // Asteroid explosion
             }
           }
         }
@@ -661,7 +663,7 @@ class SpaceScene extends Scene {
             continue;
           }
           settings.shake.onHit(this.app);
-          window.sampler("a4, 0.5");
+          window.sampler("a0, 0.5");
           const pe = this.player.e;
           this.player.e -= b.e;
           b.e -= pe;
@@ -695,7 +697,7 @@ class SpaceScene extends Scene {
               },
             });
             this.debrisList.push(debris);
-            window.sampler("a6", 1.9); // Self explosion
+            window.sampler("f1", 1.9); // Self explosion
             if (
               b.kind === "kPhotonTorpedo" ||
               b.kind === "kGaussCannonBullet"
@@ -733,11 +735,11 @@ class SpaceScene extends Scene {
             collisioning < 0 &&
             collisioning > -150
           ) {
-            // TODO or idea: kClose for flying close to asteroids
             triggerTextEffect("kClose", b.pos.x, b.pos.y, this.scale);
             b.minDistance[o._id] = -1; // Stop the loop
           }
           if (collisioning == 1) {
+            window.sampler(`a1`, 0.1); // Other impact
             const oe = o.e;
             o.showHit = performance.now() + settings.showHitMs;
             if (b.shooter?.human && b.kind !== "kEmpBlast") {
@@ -803,7 +805,7 @@ class SpaceScene extends Scene {
                   y: (2 * b.vel.y) / bnv - (0.4 * o.vel.y) / onv,
                 },
               });
-              window.sampler("a7", 1.9); // Other explosion
+              window.sampler("f0", 1.9); // Other explosion
               this.debrisList.push(debris);
               if (this.otherShips.length === 1) {
                 // This was the last ship
@@ -865,7 +867,7 @@ class SpaceScene extends Scene {
         a.e = -1;
         a.minDistance[this.player._id] = Infinity;
         this.player.killedBy = { id: "kAsteroid" };
-        window.sampler("a6", 1.9); // Self explosion
+        window.sampler("f1", 1.9); // Self explosion
 
         this.player.lives -= 1;
         const pnv = sqnorm(this.player.vel.x, this.player.vel.y) + 0.01;
@@ -879,7 +881,8 @@ class SpaceScene extends Scene {
         this.debrisList.push(debris);
         this.player.e = -1;
         this.flameList.push(...this.player.flameList);
-        window.sampler("a5", 1.9); // Asteroid explosion
+        const r = Math.floor(Math.random() * 4);
+        window.sampler(`g${r}`, 1.9); // Asteroid explosion
         newAsteroids.push(...a.split(this.player.vel, this.debrisList));
       }
       for (let o of this.otherShips) {
@@ -894,8 +897,9 @@ class SpaceScene extends Scene {
           this.debrisList.push(debris);
           o.e = -1;
           newAsteroids.push(...a.split(o.vel, this.debrisList));
-          window.sampler("a5", 1.9); // Asteroid explosion
-          window.sampler("a7", 1.9); // Other explosion
+          const r = Math.floor(Math.random() * 4);
+          window.sampler(`g${r}`, 1.9); // Asteroid explosion
+          window.sampler("f0", 1.9); // Other explosion
         }
       }
     }